Make openhands-api the canonical V1 skill
Remove the legacy V0 OpenHands API skill, move the supported V1 Cloud API skill into skills/openhands-api, update marketplace metadata and tests, and keep compatibility aliases for V1 class names.
