--- skills/openhands-api-v1/references/example_prompt.md
@@ -1,12 +1,0 @@
-# Example prompt file (OpenHands V1)
-
-Replace the text below with the task you want OpenHands to perform.
-
----
-
-Please:
-1) Explain what this repository does.
-2) List the existing test/lint commands you can find.
-3) Propose the smallest possible change to address the issue described in the task.
-
-(Do not run long loops or poll APIs. Keep external calls minimal.)
