+++ skills/openhands-api/scripts/openhands_api.ts
@@ -1,77 +1,53 @@
 /*
-OpenHands API (legacy V0) minimal client.
-
-This file mirrors the Python client in scripts/openhands_api.py.
+OpenHands Cloud API (V1) minimal client.
 
 Audience: AI agents.
 
-Recommended workflow:
-1) POST /api/conversations (create)
-2) GET /api/conversations/:id (status)
-3) GET /api/conversations/:id/events (incremental events)
-4) GET /api/conversations/:id/trajectory (full history)
+App server (Cloud):
+- Base: https://app.all-hands.dev
+- Prefix: /api/v1
+- Auth: Authorization: Bearer <OPENHANDS_CLOUD_API_KEY>
 
-Docs: https://docs.openhands.dev/api-reference/new-conversation
-Server routes: https://github.com/OpenHands/OpenHands/tree/main/openhands/server/routes
+Agent server (sandbox runtime):
+- Base: {agent_server_url}/api
+- Auth: X-Session-API-Key: <session_api_key>
+
+This is intentionally small and keeps responses mostly untyped (unknown/record)
+so it is easy to adapt.
 */
 
-export type OpenHandsAPIOptions = {
+export type OpenHandsOptions = {
   apiKey: string;
   baseUrl?: string;
 };
 
-export type CreateConversationRequest = {
-  initialUserMsg: string;
-  repository?: string;
-  selectedBranch?: string;
-  gitProvider?: string;
-  conversationInstructions?: string;
-};
+const AGENT_EVENTS_SEARCH_MAX_LIMIT = 100;
 
-export type CreateConversationResponse = {
-  status: string;
-  conversation_id: string;
-  message?: string;
-  url?: string;
-  [k: string]: unknown;
-};
-
-export type GetConversationResponse = {
-  conversation_id: string;
-  status: string;
-  url?: string;
-  [k: string]: unknown;
-};
-
-export type GetEventsResponse = {
-  events: unknown[];
-  has_more?: boolean;
-  [k: string]: unknown;
-};
-
-export type ListConversationsResponse = {
-  results: unknown[];
-  next_page_id?: string | null;
-  [k: string]: unknown;
-};
 
 export class OpenHandsAPI {
   private readonly apiKey: string;
   private readonly baseUrl: string;
 
-  constructor(opts: OpenHandsAPIOptions) {
+  constructor(opts: OpenHandsOptions) {
     if (!opts.apiKey) throw new Error("Missing apiKey");
     this.apiKey = opts.apiKey;
     this.baseUrl = (opts.baseUrl ?? "https://app.all-hands.dev").replace(/\/$/, "");
   }
 
-  private async request<T>(path: string, init?: RequestInit): Promise<T> {
-    const url = `${this.baseUrl}${path}`;
+  private get apiV1Url(): string {
+    return `${this.baseUrl}/api/v1`;
+  }
+
+  private async baseRequest<T>(
+    url: string,
+    init: RequestInit | undefined,
+    headers: Record<string, string>,
+    parseAs: "json" | "blob" = "json",
+  ): Promise<T> {
     const res = await fetch(url, {
       ...init,
       headers: {
-        Authorization: `Bearer ${this.apiKey}`,
-        "Content-Type": "application/json",
+        ...headers,
         ...(init?.headers ?? {}),
       },
     });
@@ -81,128 +57,196 @@ export class OpenHandsAPI {
       throw new Error(`OpenHands API error ${res.status} ${res.statusText}: ${text}`);
     }
 
+    if (parseAs === "blob") return (await res.blob()) as unknown as T;
     return (await res.json()) as T;
   }
 
-  async createConversation(req: CreateConversationRequest): Promise<CreateConversationResponse> {
-    return await this.request<CreateConversationResponse>("/api/conversations", {
-      method: "POST",
-      body: JSON.stringify({
-        initial_user_msg: req.initialUserMsg,
-        repository: req.repository,
-        selected_branch: req.selectedBranch,
-        git_provider: req.gitProvider,
-        conversation_instructions: req.conversationInstructions,
-      }),
-    });
-  }
-
-  async getConversation(conversationId: string): Promise<GetConversationResponse> {
-    return await this.request<GetConversationResponse>(
-      `/api/conversations/${encodeURIComponent(conversationId)}`,
-      { method: "GET" },
+  private async request<T>(
+    url: string,
+    init?: RequestInit,
+    parseAs: "json" | "blob" = "json",
+  ): Promise<T> {
+    return await this.baseRequest(
+      url,
+      init,
+      { Authorization: `Bearer ${this.apiKey}`, "Content-Type": "application/json" },
+      parseAs,
     );
   }
 
-  async getEvents(
-    conversationId: string,
-    opts?: { startId?: number; endId?: number; reverse?: boolean; limit?: number },
-  ): Promise<GetEventsResponse> {
-    const limit = Math.max(1, Math.min(100, opts?.limit ?? 20));
-    const params = new URLSearchParams({
-      start_id: String(opts?.startId ?? 0),
-      reverse: String(Boolean(opts?.reverse)),
-      limit: String(limit),
-    });
-    if (opts?.endId !== undefined) params.set("end_id", String(opts.endId));
+  // -----------------------------
+  // App server endpoints
+  // -----------------------------
 
-    return await this.request<GetEventsResponse>(
-      `/api/conversations/${encodeURIComponent(conversationId)}/events?${params.toString()}`,
-      { method: "GET" },
-    );
+  async usersMe(): Promise<Record<string, unknown>> {
+    return await this.request(`${this.apiV1Url}/users/me`, { method: "GET" });
   }
 
-  async getTrajectory(conversationId: string): Promise<Record<string, unknown>> {
-    return await this.request<Record<string, unknown>>(
-      `/api/conversations/${encodeURIComponent(conversationId)}/trajectory`,
-      { method: "GET" },
-    );
-  }
-
-  async listConversations(opts?: {
-    limit?: number;
-    pageId?: string;
-    selectedRepository?: string;
-    includeSubConversations?: boolean;
-  }): Promise<ListConversationsResponse> {
-    const params = new URLSearchParams({
-      limit: String(Math.max(1, opts?.limit ?? 20)),
-    });
-    if (opts?.pageId) params.set("page_id", opts.pageId);
-    if (opts?.selectedRepository) params.set("selected_repository", opts.selectedRepository);
-    if (opts?.includeSubConversations !== undefined) {
-      params.set("include_sub_conversations", String(Boolean(opts.includeSubConversations)));
-    }
-
-    return await this.request<ListConversationsResponse>(`/api/conversations?${params.toString()}`, {
+  async appConversationsSearch(limit = 20): Promise<Record<string, unknown>> {
+    const safeLimit = Number.isFinite(limit) ? Math.trunc(limit) : 1;
+    const params = new URLSearchParams({ limit: String(Math.max(1, safeLimit)) });
+    return await this.request(`${this.apiV1Url}/app-conversations/search?${params.toString()}`, {
       method: "GET",
     });
   }
 
-  async updateConversationTitle(conversationId: string, title: string): Promise<Record<string, unknown>> {
-    return await this.request<Record<string, unknown>>(
-      `/api/conversations/${encodeURIComponent(conversationId)}`,
-      { method: "PATCH", body: JSON.stringify({ title }) },
-    );
-  }
-
-  async deleteConversation(conversationId: string): Promise<Record<string, unknown>> {
-    return await this.request<Record<string, unknown>>(
-      `/api/conversations/${encodeURIComponent(conversationId)}`,
-      { method: "DELETE" },
-    );
-  }
-
-  async addMessage(conversationId: string, message: string): Promise<Record<string, unknown>> {
-    return await this.request<Record<string, unknown>>(
-      `/api/conversations/${encodeURIComponent(conversationId)}/message`,
-      { method: "POST", body: JSON.stringify({ message }) },
-    );
-  }
-
-  async listWorkspaceFiles(conversationId: string, path?: string): Promise<unknown> {
+  async appConversationsGetBatch(ids: string[]): Promise<Array<Record<string, unknown>>> {
+    if (ids.length === 0) return [];
     const params = new URLSearchParams();
-    if (path !== undefined) params.set("path", path);
-    const qs = params.toString();
-    return await this.request<unknown>(
-      `/api/conversations/${encodeURIComponent(conversationId)}/list-files${qs ? `?${qs}` : ""}`,
+    for (const id of ids) params.append("ids", id);
+    return await this.request(`${this.apiV1Url}/app-conversations?${params.toString()}`, {
+      method: "GET",
+    });
+  }
+
+  async conversationEventsCount(appConversationId: string): Promise<number> {
+    const res = await this.request<number>(
+      `${this.apiV1Url}/conversation/${encodeURIComponent(appConversationId)}/events/count`,
+      { method: "GET" },
+    );
+    return Number(res);
+  }
+
+  async appConversationDownloadZip(appConversationId: string): Promise<Blob> {
+    const url = `${this.apiV1Url}/app-conversations/${encodeURIComponent(appConversationId)}/download`;
+    return await this.request<Blob>(url, { method: "GET" }, "blob");
+  }
+
+  async appConversationStart(req: {
+    initialMessage: string;
+    selectedRepository?: string;
+    selectedBranch?: string;
+    title?: string;
+    run?: boolean;
+  }): Promise<Record<string, unknown>> {
+    // NOTE: In many deployments this returns a *start-task* object.
+    // `id` is usually the start_task_id; use `app_conversation_id` (if present)
+    // for `/download` and `/conversation/.../events/...` endpoints.
+    // If `app_conversation_id` is missing, fetch it via:
+    // GET /api/v1/app-conversations/start-tasks?ids=<start_task_id>
+
+    const payload: Record<string, unknown> = {
+      initial_message: {
+        role: "user",
+        // V1 expects `content` as an array of parts, even for a single text message.
+        content: [{ type: "text", text: req.initialMessage }],
+        run: req.run ?? true,
+      },
+    };
+    if (req.selectedRepository) payload.selected_repository = req.selectedRepository;
+    if (req.selectedBranch) payload.selected_branch = req.selectedBranch;
+    if (req.title) payload.title = req.title;
+
+    return await this.request(`${this.apiV1Url}/app-conversations`, {
+      method: "POST",
+      body: JSON.stringify(payload),
+    });
+  }
+
+  async appConversationsStartTasksGetBatch(ids: string[]): Promise<Array<Record<string, unknown>>> {
+    if (ids.length === 0) return [];
+    const params = new URLSearchParams();
+    for (const id of ids) params.append("ids", id);
+    return await this.request(`${this.apiV1Url}/app-conversations/start-tasks?${params.toString()}`, {
+      method: "GET",
+    });
+  }
+
+  // -----------------------------
+  // Agent server endpoints
+  // -----------------------------
+
+  private async agentRequest<T>(
+    agentServerUrl: string,
+    sessionApiKey: string,
+    path: string,
+    init?: RequestInit,
+  ): Promise<T> {
+    const base = agentServerUrl.replace(/\/$/, "");
+    const url = `${base}${path}`;
+    return await this.baseRequest(
+      url,
+      init,
+      { "X-Session-API-Key": sessionApiKey, "Content-Type": "application/json" },
+      "json",
+    );
+  }
+
+  private buildAgentEventFilterParams(opts?: {
+    timestampGte?: string;
+    timestampLt?: string;
+    kind?: string;
+    source?: string;
+    body?: string;
+  }): URLSearchParams {
+    const params = new URLSearchParams();
+    if (opts?.timestampGte) params.set("timestamp__gte", opts.timestampGte);
+    if (opts?.timestampLt) params.set("timestamp__lt", opts.timestampLt);
+    if (opts?.kind) params.set("kind", opts.kind);
+    if (opts?.source) params.set("source", opts.source);
+    if (opts?.body) params.set("body", opts.body);
+    return params;
+  }
+
+  async agentEventsCount(agentServerUrl: string, sessionApiKey: string, conversationId: string, opts?: {
+    timestampGte?: string;
+    timestampLt?: string;
+    kind?: string;
+    source?: string;
+    body?: string;
+  }): Promise<number> {
+    const qs = this.buildAgentEventFilterParams(opts).toString();
+    const suffix = qs ? `?${qs}` : "";
+
+    const n = await this.agentRequest<number>(
+      agentServerUrl,
+      sessionApiKey,
+      `/api/conversations/${encodeURIComponent(conversationId)}/events/count${suffix}`,
+      { method: "GET" },
+    );
+    return Number(n);
+  }
+
+  async agentEventsSearch(agentServerUrl: string, sessionApiKey: string, conversationId: string, opts?: {
+    limit?: number;
+    sortOrder?: "TIMESTAMP" | "TIMESTAMP_DESC";
+    timestampGte?: string;
+    timestampLt?: string;
+    kind?: string;
+    source?: string;
+    body?: string;
+  }): Promise<Record<string, unknown>> {
+    const params = this.buildAgentEventFilterParams(opts);
+
+    // Cap limit to keep responses small and consistent across clients.
+    const rawLimit = opts?.limit ?? 50;
+    const safeLimit = Number.isFinite(rawLimit) ? Math.trunc(rawLimit) : 1;
+    const limit = Math.max(1, Math.min(AGENT_EVENTS_SEARCH_MAX_LIMIT, safeLimit));
+    params.set("limit", String(limit));
+
+    if (opts?.sortOrder) params.set("sort_order", opts.sortOrder);
+
+    return await this.agentRequest<Record<string, unknown>>(
+      agentServerUrl,
+      sessionApiKey,
+      `/api/conversations/${encodeURIComponent(conversationId)}/events/search?${params.toString()}`,
       { method: "GET" },
     );
   }
 
-  async getFileContent(conversationId: string, filePath: string): Promise<Record<string, unknown>> {
-    const params = new URLSearchParams({ file: filePath });
-    return await this.request<Record<string, unknown>>(
-      `/api/conversations/${encodeURIComponent(conversationId)}/select-file?${params.toString()}`,
-      { method: "GET" },
+  async agentExecuteBash(agentServerUrl: string, sessionApiKey: string, command: string, cwd?: string): Promise<Record<string, unknown>> {
+    const payload: Record<string, unknown> = { command, timeout: 30 };
+    if (cwd) payload.cwd = cwd;
+
+    return await this.agentRequest<Record<string, unknown>>(
+      agentServerUrl,
+      sessionApiKey,
+      `/api/bash/execute_bash_command`,
+      { method: "POST", body: JSON.stringify(payload) },
     );
-  }
-
-  async pollUntilTerminal(
-    conversationId: string,
-    opts?: { timeoutMs?: number; pollIntervalMs?: number },
-  ): Promise<GetConversationResponse> {
-    const timeoutMs = opts?.timeoutMs ?? 20 * 60_000;
-    const pollIntervalMs = opts?.pollIntervalMs ?? 30_000;
-
-    const start = Date.now();
-    while (Date.now() - start < timeoutMs) {
-      const convo = await this.getConversation(conversationId);
-      const s = String(convo.status ?? "").toUpperCase();
-      if (["STOPPED", "ERROR", "FAILED", "CANCELLED"].includes(s)) return convo;
-      await new Promise((r) => setTimeout(r, pollIntervalMs));
-    }
-
-    throw new Error(`Conversation ${conversationId} did not reach terminal within ${timeoutMs}ms`);
   }
 }
+
+
+export type OpenHandsV1Options = OpenHandsOptions;
+export { OpenHandsAPI as OpenHandsV1API };
